LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/constructor/004-expected.txt

CONSOLE MESSAGE: Wrong protocol for WebSocket ''
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0001b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0002b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0003b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0004b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0005b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0006b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0007b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0008b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0009b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u000Ab'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u000Bb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u000Cb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u000Db'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u000Eb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u000Fb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0010b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0011b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0012b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0013b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0014b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0015b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0016b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0017b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0018b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0019b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u001Ab'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u001Bb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u001Cb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u001Db'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u001Eb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u001Fb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u007Fb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0080b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0081b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0082b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0083b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0084b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0085b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0086b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0087b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0088b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0089b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u008Ab'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u008Bb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u008Cb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u008Db'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u008Eb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u008Fb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0090b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0091b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0092b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0093b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0094b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0095b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0096b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0097b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0098b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0099b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u009Ab'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u009Bb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u009Cb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u009Db'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u009Eb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u009Fb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00A0b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00A1b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00A2b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00A3b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00A4b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00A5b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00A6b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00A7b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00A8b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00A9b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00AAb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00ABb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00ACb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00ADb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00AEb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00AFb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00B0b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00B1b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00B2b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00B3b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00B4b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00B5b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00B6b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00B7b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00B8b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00B9b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00BAb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00BBb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00BCb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00BDb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00BEb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00BFb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00C0b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00C1b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00C2b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00C3b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00C4b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00C5b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00C6b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00C7b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00C8b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00C9b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00CAb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00CBb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00CCb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00CDb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00CEb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00CFb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00D0b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00D1b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00D2b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00D3b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00D4b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00D5b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00D6b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00D7b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00D8b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00D9b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00DAb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00DBb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00DCb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00DDb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00DEb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00DFb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00E0b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00E1b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00E2b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00E3b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00E4b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00E5b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00E6b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00E7b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00E8b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00E9b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00EAb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00EBb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00ECb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00EDb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00EEb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00EFb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00F0b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00F1b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00F2b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00F3b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00F4b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00F5b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00F6b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00F7b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00F8b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00F9b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00FAb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00FBb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00FCb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00FDb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00FEb'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u00FFb'

PASS WebSockets: new WebSocket(url, invalid protocol)
PASS WebSockets: new WebSocket(url, invalid protocol) 1
PASS WebSockets: new WebSocket(url, invalid protocol) 2
PASS WebSockets: new WebSocket(url, invalid protocol) 3
PASS WebSockets: new WebSocket(url, invalid protocol) 4
PASS WebSockets: new WebSocket(url, invalid protocol) 5
PASS WebSockets: new WebSocket(url, invalid protocol) 6
PASS WebSockets: new WebSocket(url, invalid protocol) 7
PASS WebSockets: new WebSocket(url, invalid protocol) 8
PASS WebSockets: new WebSocket(url, invalid protocol) 9
PASS WebSockets: new WebSocket(url, invalid protocol) 10
PASS WebSockets: new WebSocket(url, invalid protocol) 11
PASS WebSockets: new WebSocket(url, invalid protocol) 12
PASS WebSockets: new WebSocket(url, invalid protocol) 13
PASS WebSockets: new WebSocket(url, invalid protocol) 14
PASS WebSockets: new WebSocket(url, invalid protocol) 15
PASS WebSockets: new WebSocket(url, invalid protocol) 16
PASS WebSockets: new WebSocket(url, invalid protocol) 17
PASS WebSockets: new WebSocket(url, invalid protocol) 18
PASS WebSockets: new WebSocket(url, invalid protocol) 19
PASS WebSockets: new WebSocket(url, invalid protocol) 20
PASS WebSockets: new WebSocket(url, invalid protocol) 21
PASS WebSockets: new WebSocket(url, invalid protocol) 22
PASS WebSockets: new WebSocket(url, invalid protocol) 23
PASS WebSockets: new WebSocket(url, invalid protocol) 24
PASS WebSockets: new WebSocket(url, invalid protocol) 25
PASS WebSockets: new WebSocket(url, invalid protocol) 26
PASS WebSockets: new WebSocket(url, invalid protocol) 27
PASS WebSockets: new WebSocket(url, invalid protocol) 28
PASS WebSockets: new WebSocket(url, invalid protocol) 29
PASS WebSockets: new WebSocket(url, invalid protocol) 30
PASS WebSockets: new WebSocket(url, invalid protocol) 31
PASS WebSockets: new WebSocket(url, invalid protocol) 32
PASS WebSockets: new WebSocket(url, invalid protocol) 33
PASS WebSockets: new WebSocket(url, invalid protocol) 34
PASS WebSockets: new WebSocket(url, invalid protocol) 35
PASS WebSockets: new WebSocket(url, invalid protocol) 36
PASS WebSockets: new WebSocket(url, invalid protocol) 37
PASS WebSockets: new WebSocket(url, invalid protocol) 38
PASS WebSockets: new WebSocket(url, invalid protocol) 39
PASS WebSockets: new WebSocket(url, invalid protocol) 40
PASS WebSockets: new WebSocket(url, invalid protocol) 41
PASS WebSockets: new WebSocket(url, invalid protocol) 42
PASS WebSockets: new WebSocket(url, invalid protocol) 43
PASS WebSockets: new WebSocket(url, invalid protocol) 44
PASS WebSockets: new WebSocket(url, invalid protocol) 45
PASS WebSockets: new WebSocket(url, invalid protocol) 46
PASS WebSockets: new WebSocket(url, invalid protocol) 47
PASS WebSockets: new WebSocket(url, invalid protocol) 48
PASS WebSockets: new WebSocket(url, invalid protocol) 49
PASS WebSockets: new WebSocket(url, invalid protocol) 50
PASS WebSockets: new WebSocket(url, invalid protocol) 51
PASS WebSockets: new WebSocket(url, invalid protocol) 52
PASS WebSockets: new WebSocket(url, invalid protocol) 53
PASS WebSockets: new WebSocket(url, invalid protocol) 54
PASS WebSockets: new WebSocket(url, invalid protocol) 55
PASS WebSockets: new WebSocket(url, invalid protocol) 56
PASS WebSockets: new WebSocket(url, invalid protocol) 57
PASS WebSockets: new WebSocket(url, invalid protocol) 58
PASS WebSockets: new WebSocket(url, invalid protocol) 59
PASS WebSockets: new WebSocket(url, invalid protocol) 60
PASS WebSockets: new WebSocket(url, invalid protocol) 61
PASS WebSockets: new WebSocket(url, invalid protocol) 62
PASS WebSockets: new WebSocket(url, invalid protocol) 63
PASS WebSockets: new WebSocket(url, invalid protocol) 64
PASS WebSockets: new WebSocket(url, invalid protocol) 65
PASS WebSockets: new WebSocket(url, invalid protocol) 66
PASS WebSockets: new WebSocket(url, invalid protocol) 67
PASS WebSockets: new WebSocket(url, invalid protocol) 68
PASS WebSockets: new WebSocket(url, invalid protocol) 69
PASS WebSockets: new WebSocket(url, invalid protocol) 70
PASS WebSockets: new WebSocket(url, invalid protocol) 71
PASS WebSockets: new WebSocket(url, invalid protocol) 72
PASS WebSockets: new WebSocket(url, invalid protocol) 73
PASS WebSockets: new WebSocket(url, invalid protocol) 74
PASS WebSockets: new WebSocket(url, invalid protocol) 75
PASS WebSockets: new WebSocket(url, invalid protocol) 76
PASS WebSockets: new WebSocket(url, invalid protocol) 77
PASS WebSockets: new WebSocket(url, invalid protocol) 78
PASS WebSockets: new WebSocket(url, invalid protocol) 79
PASS WebSockets: new WebSocket(url, invalid protocol) 80
PASS WebSockets: new WebSocket(url, invalid protocol) 81
PASS WebSockets: new WebSocket(url, invalid protocol) 82
PASS WebSockets: new WebSocket(url, invalid protocol) 83
PASS WebSockets: new WebSocket(url, invalid protocol) 84
PASS WebSockets: new WebSocket(url, invalid protocol) 85
PASS WebSockets: new WebSocket(url, invalid protocol) 86
PASS WebSockets: new WebSocket(url, invalid protocol) 87
PASS WebSockets: new WebSocket(url, invalid protocol) 88
PASS WebSockets: new WebSocket(url, invalid protocol) 89
PASS WebSockets: new WebSocket(url, invalid protocol) 90
PASS WebSockets: new WebSocket(url, invalid protocol) 91
PASS WebSockets: new WebSocket(url, invalid protocol) 92
PASS WebSockets: new WebSocket(url, invalid protocol) 93
PASS WebSockets: new WebSocket(url, invalid protocol) 94
PASS WebSockets: new WebSocket(url, invalid protocol) 95
PASS WebSockets: new WebSocket(url, invalid protocol) 96
PASS WebSockets: new WebSocket(url, invalid protocol) 97
PASS WebSockets: new WebSocket(url, invalid protocol) 98
PASS WebSockets: new WebSocket(url, invalid protocol) 99
PASS WebSockets: new WebSocket(url, invalid protocol) 100
PASS WebSockets: new WebSocket(url, invalid protocol) 101
PASS WebSockets: new WebSocket(url, invalid protocol) 102
PASS WebSockets: new WebSocket(url, invalid protocol) 103
PASS WebSockets: new WebSocket(url, invalid protocol) 104
PASS WebSockets: new WebSocket(url, invalid protocol) 105
PASS WebSockets: new WebSocket(url, invalid protocol) 106
PASS WebSockets: new WebSocket(url, invalid protocol) 107
PASS WebSockets: new WebSocket(url, invalid protocol) 108
PASS WebSockets: new WebSocket(url, invalid protocol) 109
PASS WebSockets: new WebSocket(url, invalid protocol) 110
PASS WebSockets: new WebSocket(url, invalid protocol) 111
PASS WebSockets: new WebSocket(url, invalid protocol) 112
PASS WebSockets: new WebSocket(url, invalid protocol) 113
PASS WebSockets: new WebSocket(url, invalid protocol) 114
PASS WebSockets: new WebSocket(url, invalid protocol) 115
PASS WebSockets: new WebSocket(url, invalid protocol) 116
PASS WebSockets: new WebSocket(url, invalid protocol) 117
PASS WebSockets: new WebSocket(url, invalid protocol) 118
PASS WebSockets: new WebSocket(url, invalid protocol) 119
PASS WebSockets: new WebSocket(url, invalid protocol) 120
PASS WebSockets: new WebSocket(url, invalid protocol) 121
PASS WebSockets: new WebSocket(url, invalid protocol) 122
PASS WebSockets: new WebSocket(url, invalid protocol) 123
PASS WebSockets: new WebSocket(url, invalid protocol) 124
PASS WebSockets: new WebSocket(url, invalid protocol) 125
PASS WebSockets: new WebSocket(url, invalid protocol) 126
PASS WebSockets: new WebSocket(url, invalid protocol) 127
PASS WebSockets: new WebSocket(url, invalid protocol) 128
PASS WebSockets: new WebSocket(url, invalid protocol) 129
PASS WebSockets: new WebSocket(url, invalid protocol) 130
PASS WebSockets: new WebSocket(url, invalid protocol) 131
PASS WebSockets: new WebSocket(url, invalid protocol) 132
PASS WebSockets: new WebSocket(url, invalid protocol) 133
PASS WebSockets: new WebSocket(url, invalid protocol) 134
PASS WebSockets: new WebSocket(url, invalid protocol) 135
PASS WebSockets: new WebSocket(url, invalid protocol) 136
PASS WebSockets: new WebSocket(url, invalid protocol) 137
PASS WebSockets: new WebSocket(url, invalid protocol) 138
PASS WebSockets: new WebSocket(url, invalid protocol) 139
PASS WebSockets: new WebSocket(url, invalid protocol) 140
PASS WebSockets: new WebSocket(url, invalid protocol) 141
PASS WebSockets: new WebSocket(url, invalid protocol) 142
PASS WebSockets: new WebSocket(url, invalid protocol) 143
PASS WebSockets: new WebSocket(url, invalid protocol) 144
PASS WebSockets: new WebSocket(url, invalid protocol) 145
PASS WebSockets: new WebSocket(url, invalid protocol) 146
PASS WebSockets: new WebSocket(url, invalid protocol) 147
PASS WebSockets: new WebSocket(url, invalid protocol) 148
PASS WebSockets: new WebSocket(url, invalid protocol) 149
PASS WebSockets: new WebSocket(url, invalid protocol) 150
PASS WebSockets: new WebSocket(url, invalid protocol) 151
PASS WebSockets: new WebSocket(url, invalid protocol) 152
PASS WebSockets: new WebSocket(url, invalid protocol) 153
PASS WebSockets: new WebSocket(url, invalid protocol) 154
PASS WebSockets: new WebSocket(url, invalid protocol) 155
PASS WebSockets: new WebSocket(url, invalid protocol) 156
PASS WebSockets: new WebSocket(url, invalid protocol) 157
PASS WebSockets: new WebSocket(url, invalid protocol) 158
PASS WebSockets: new WebSocket(url, invalid protocol) 159
PASS WebSockets: new WebSocket(url, invalid protocol) 160
